backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
passlib
flask-security-too
flask-cors
bleach
requests
